requirements.txt
alembic==1.13.3
amqp==5.3.1
async-timeout==5.0.1
bidict==0.23.1
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
email_validator==2.2.0
eventlet==0.39.1
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-Security==5.5.2
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.5.1
load-dotenv==0.1.0
Mako==1.3.5
MarkupSafe==2.1.5
passlib==1.7.4
prompt_toolkit==3.0.50
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
redis==5.2.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
wsproto==1.2.0
WTForms==3.1.2
